Leftover type of `provideSomeLayer` no longer required for Scala 3
