+++ core/shared/src/main/scala-2/zio/ZIOCompanionVersionSpecific.scala
@@ -4,6 +4,17 @@ import zio.ZIO.{Async, asyncInterrupt, blocking}
 
 import java.io.IOException
 
+final class ProvideSomeLayer[R0, -R, +E, +A](private val self: ZIO[R, E, A]) extends AnyVal {
+  def apply[E1 >: E, R1](
+    layer: => ZLayer[R0, E1, R1]
+  )(implicit
+    ev: R0 with R1 <:< R,
+    tagged: EnvironmentTag[R1],
+    trace: Trace
+  ): ZIO[R0, E1, A] =
+    self.asInstanceOf[ZIO[R0 with R1, E, A]].provideLayer(ZLayer.environment[R0] ++ layer)
+}
+
 trait ZIOCompanionVersionSpecific {
 
   /**
@@ -168,6 +179,9 @@ trait ZIOCompanionVersionSpecific {
           }
       }
     }
+
+  val ProvideSomeLayer = zio.ProvideSomeLayer
+  type ProvideSomeLayer[R0, -R, +E, +A] = zio.ProvideSomeLayer[R0, R, E, A]
 
   /**
    * Returns an effect that models success with the specified value.
